django_application/requirements.txt
asgiref==3.6.0 
django==4.1.7 
django-environ==0.10.0
djangorestframework==3.14.0 
pytz==2022.7.1 
sqlparse==0.4.3
Faker==18.3.1 
django-seed==0.3.1 
python-dateutil==2.8.2 
six==1.16.0 
toposort==1.10